validation/smoke-test/enhanced/test-flask-api-enhanced/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
pytest==7.4.3
pytest-flask==1.3.0
